+++ src/nginx/generate-dhparam.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 # from https://github.com/jwilder/nginx-proxy
 
 # The first argument is the bit depth of the dhparam, or 2048 if unspecified
@@ -14,9 +14,11 @@ GEN_LOCKFILE="/tmp/dhparam_generating.lock"
 PREGEN_HASH=$(md5sum $PREGEN_DHPARAM_FILE | cut -d" " -f1)
 if [[ -f $DHPARAM_FILE ]]; then
     CURRENT_HASH=$(md5sum $DHPARAM_FILE | cut -d" " -f1)
-    if [[ $PREGEN_HASH != $CURRENT_HASH ]]; then
+    if [[ "$PREGEN_HASH" != "$CURRENT_HASH" ]]; then
         # There is already a dhparam, and it's not the default
-        echo "Custom dhparam.pem file found, generation skipped"
+        if [ "${silent_init}" = false ]; then
+            echo "Custom dhparam.pem file found, generation skipped"
+        fi
         exit 0
     fi
 
@@ -26,10 +28,12 @@ if [[ -f $DHPARAM_FILE ]]; then
     fi
 fi
 
-cat >&2 <<-EOT
+if [ "${silent_init}" = false ]; then
+cat >&2 <<EOT
 INFO: A pre-generated dhparam.pem will be used for now while a new one
 is being generated in the background. Once the new dhparam.pem is in place, nginx will be reloaded.
 EOT
+fi
 
 # Put the default dhparam file in place so we can start immediately
 cp $PREGEN_DHPARAM_FILE $DHPARAM_FILE
@@ -38,9 +42,9 @@ touch $GEN_LOCKFILE
 # Generate a new dhparam in the background in a low priority and reload nginx when finished (grep removes the progress indicator).
 (
     (
-        nice -n +5 openssl dhparam -out $DHPARAM_FILE.tmp $DHPARAM_BITS 2>&1 \
+        nice -n +5 openssl dhparam -out "${DHPARAM_FILE}.tmp" "${DHPARAM_BITS}" 2>&1 \
         && mv $DHPARAM_FILE.tmp $DHPARAM_FILE \
-        && echo "dhparam generation complete, reloading nginx" \
+        && echo "INFO: New dhparam generation complete, reloading nginx..." \
         && nginx -s reload
     ) | grep -vE '^[\.+]+'
     rm $GEN_LOCKFILE